This commit was manufactured by cvs2svn to create tag 'WX_2_4_1'.

+++ src/unix/utilsx11.cpp
@@ -6,7 +6,7 @@
 // Created:     25.03.02
 // RCS-ID:      $Id$
 // Copyright:   (c) wxWindows team
-// Licence:     wxWindows licence
+// Licence:     wxWindows license
 /////////////////////////////////////////////////////////////////////////////
 
 #if defined(__WXX11__) || defined(__WXGTK__) || defined(__WXMOTIF__)
@@ -276,7 +276,7 @@ static bool wxQueryWMspecSupport(Display *display, Window rootWnd, Atom feature)
     
     // Is the WM ICCCM supporting?
     XGetWindowProperty(display, rootWnd,
-                       _NET_SUPPORTING_WM_CHECK, 0, LONG_MAX,
+                       _NET_SUPPORTING_WM_CHECK, 0, UINT_MAX,
                        False, XA_WINDOW, &type, &format, &nwins,
                        &after, (unsigned char **)&wins);
     if ( type != XA_WINDOW || nwins <= 0 || wins[0] == None )
@@ -285,7 +285,7 @@ static bool wxQueryWMspecSupport(Display *display, Window rootWnd, Atom feature)
 
     // Query for supported features:
     XGetWindowProperty(display, rootWnd,
-                       _NET_SUPPORTED, 0, LONG_MAX,
+                       _NET_SUPPORTED, 0, UINT_MAX,
                        False, XA_ATOM, &type, &format, &natoms,
                        &after, (unsigned char **)&atoms);
     if ( type != XA_ATOM || atoms == NULL )
